refactor(drawer): refactors to fixes some issues
