+++ src/components/KeyResult/Single/Drawer/Body/index.ts
@@ -1,0 +1,3 @@
+import KeyResultDrawerBody from './body'
+
+export default KeyResultDrawerBody
